+++ src/main/java/io/tickerstorm/data/StooqFileConverter.java
@@ -133,7 +133,7 @@ public class StooqFileConverter extends BaseFileConverter implements DataConvert
   @Override
   public void onDirectoryChange(File file) {
 
-    if (file.getPath().contains(provider())) {
+    if (file.getPath().contains(provider()) && !file.getName().equalsIgnoreCase(provider())) {
       if (file.isDirectory() && file.list().length == 0) {
         logger.info("Deleting " + file.getPath() + " since it's empty");
         FileUtils.deleteQuietly(file);
